+++ ParcelInvoiceMaker/ParcelInvoiceMaker/ParcelProcessor/ParcelProcessor.swift
@@ -6,45 +6,6 @@
 
 import Foundation
 
-enum Discount: Int {
-    case none = 0,
-         vip,
-         coupon
-    
-    var strategy: DiscountStrategy {
-        switch self {
-        case .none:
-            return NoDiscount()
-        case .vip:
-            return VIPDiscount()
-        case .coupon:
-            return CouponDiscount()
-        }
-    }
-}
-
-protocol DiscountStrategy {
-    func applyDiscount(deliveryCost: Int) -> Int
-}
-
-struct NoDiscount: DiscountStrategy {
-    func applyDiscount(deliveryCost: Int) -> Int {
-        return deliveryCost
-    }
-}
-
-struct VIPDiscount: DiscountStrategy {
-    func applyDiscount(deliveryCost: Int) -> Int {
-        return deliveryCost / 5 * 4
-    }
-}
-
-struct CouponDiscount: DiscountStrategy {
-    func applyDiscount(deliveryCost: Int) -> Int {
-        return deliveryCost / 2
-    }
-}
-
 class ParcelInformation {
     let receiver: Receiver
     let deliveryCost: Cost
@@ -75,10 +36,6 @@ class ParcelOrderProcessor {
         
         onComplete(parcelInformation)
     }
-}
-
-protocol ParcelInformationPersistence {
-    func save(parcelInformation: ParcelInformation)
 }
 
 class DatabaseParcelInformationPersistence: ParcelInformationPersistence {
